+++ certbot/CHANGELOG.md
@@ -22,6 +22,7 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 
 * The Certbot snap no longer loads packages installed via `pip install --user`. This
   was unintended and DNS plugins should be installed via `snap` instead.
+* `certbot-dns-google` would sometimes crash with HTTP 409/412 errors when used with very large zones (#6036)
 
 More details about these changes can be found on our GitHub repo.